+++ kyuubi-hive-jdbc/src/main/resources/META-INF/services/java.sql.Driver
@@ -14,3 +14,4 @@
 # limitations under the License.
 
 org.apache.kyuubi.jdbc.KyuubiDriver
+org.apache.kyuubi.jdbc.KyuubiHiveDriver
